+++ crates/mesh-llm-host-runtime/src/inference/skippy/resolver.rs
@@ -6,6 +6,12 @@ mod translation;
 mod types;
 
 #[cfg(test)]
+mod test_support;
+
+#[cfg(test)]
+mod native_mtp_tests;
+
+#[cfg(test)]
 mod tests;
 
 pub(crate) use resolution::resolve_skippy_config;
