+++ crates/mesh-llm-host-runtime/src/inference/skippy/resolver/tests.rs
@@ -1,88 +1,17 @@
+use super::test_support::*;
 use super::*;
-use crate::inference::skippy::{SkippyPackageIdentity, SkippyTelemetryOptions, StageWireDType};
+use crate::inference::skippy::{SkippyTelemetryOptions, StageWireDType};
 use crate::plugin::{MeshConfig, ReasoningBudget, RequestDefaultsConfig};
 use serde_json::Value;
 use skippy_protocol::{LoadMode, StageKvCacheMode, StageKvCachePayload};
-use skippy_runtime::package::{
-    PackageGenerationInfo, PackageSpeculativeDecodingInfo, PackageSpeculativeStrategyInfo,
-    PackageWindowPolicyInfo,
-};
 use skippy_server::{EmbeddedReasoningEnabled, EmbeddedReasoningFormat};
-use std::{
-    collections::BTreeMap,
-    io::Write,
-    path::{Path, PathBuf},
-};
+use std::path::Path;
 use tempfile::NamedTempFile;
 
 const FULL_SURFACE_VALID_FIXTURE: &str =
     include_str!("../../../../tests/fixtures/skippy_full_surface_valid.toml");
 const FULL_SURFACE_INVALID_FIXTURE: &str =
     include_str!("../../../../tests/fixtures/skippy_full_surface_invalid.toml");
-
-fn fake_package_identity(layer_count: u32) -> SkippyPackageIdentity {
-    SkippyPackageIdentity {
-        package_ref: "gguf:///models/qwen.gguf".to_string(),
-        manifest_sha256: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-            .to_string(),
-        source_model_path: PathBuf::from("/models/qwen.gguf"),
-        source_model_sha256: "abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789"
-            .to_string(),
-        source_model_bytes: 1234,
-        source_files: Vec::new(),
-        layer_count,
-        activation_width: 4096,
-        tensor_count: 100,
-        generation: None,
-    }
-}
-
-fn fake_hf_package_identity(layer_count: u32) -> SkippyPackageIdentity {
-    let mut package = fake_package_identity(layer_count);
-    package.package_ref = "hf://meshllm/Qwen3-8B-Q4_K_M-layers".to_string();
-    package
-}
-
-fn parse_config(toml: &str) -> MeshConfig {
-    toml::from_str(toml).expect("config should parse")
-}
-
-fn push_gguf_string(bytes: &mut Vec<u8>, value: &str) {
-    bytes.extend_from_slice(&(value.len() as u64).to_le_bytes());
-    bytes.extend_from_slice(value.as_bytes());
-}
-
-fn push_u32_kv(bytes: &mut Vec<u8>, key: &str, value: u32) {
-    push_gguf_string(bytes, key);
-    bytes.extend_from_slice(&4u32.to_le_bytes());
-    bytes.extend_from_slice(&value.to_le_bytes());
-}
-
-fn push_string_kv(bytes: &mut Vec<u8>, key: &str, value: &str) {
-    push_gguf_string(bytes, key);
-    bytes.extend_from_slice(&8u32.to_le_bytes());
-    push_gguf_string(bytes, value);
-}
-
-fn temp_model_file() -> NamedTempFile {
-    let mut file = NamedTempFile::new().expect("temp model file");
-    let mut bytes = Vec::new();
-    bytes.extend_from_slice(b"GGUF");
-    bytes.extend_from_slice(&2u32.to_le_bytes());
-    bytes.extend_from_slice(&0i64.to_le_bytes());
-    bytes.extend_from_slice(&8i64.to_le_bytes());
-    push_string_kv(&mut bytes, "general.architecture", "llama");
-    push_string_kv(&mut bytes, "tokenizer.ggml.model", "gpt2");
-    push_u32_kv(&mut bytes, "llama.context_length", 8192);
-    push_u32_kv(&mut bytes, "llama.embedding_length", 4096);
-    push_u32_kv(&mut bytes, "llama.block_count", 24);
-    push_u32_kv(&mut bytes, "llama.attention.head_count", 32);
-    push_u32_kv(&mut bytes, "llama.attention.head_count_kv", 8);
-    push_u32_kv(&mut bytes, "llama.attention.key_length", 128);
-    file.write_all(&bytes).expect("write fake gguf");
-    file.flush().expect("flush fake gguf");
-    file
-}
 
 fn resolve_qwen_config_with_request_defaults(
     mesh_config: &MeshConfig,
@@ -99,31 +28,6 @@ fn resolve_qwen_config_with_request_defaults(
         package_generation: None,
     })
     .expect("qwen config should resolve")
-}
-
-fn native_mtp_generation() -> PackageGenerationInfo {
-    let mut strategies = BTreeMap::new();
-    strategies.insert(
-        "native-mtp-n1".to_string(),
-        PackageSpeculativeStrategyInfo {
-            strategy_type: "native-mtp".to_string(),
-            prediction_depth: Some(1),
-            layer_indices: vec![46],
-            window_policy: Some(PackageWindowPolicyInfo {
-                default: "fixed".to_string(),
-                initial_window: 1,
-                min_window: 1,
-                max_window: 1,
-            }),
-        },
-    );
-
-    PackageGenerationInfo {
-        speculative_decoding: Some(PackageSpeculativeDecodingInfo {
-            default: "native-mtp-n1".to_string(),
-            strategies,
-        }),
-    }
 }
 
 fn assert_request_override_keeps_load_time_config(
@@ -818,136 +722,6 @@ draft_selection_policy = "auto"
     assert_eq!(resolved.speculative.mode, "disabled");
     assert!(resolved.speculative.draft_model_path.is_none());
     assert!(!resolved.speculative.explicit);
-}
-
-#[test]
-fn speculative_strategy_auto_without_package_generation_disables_native_mtp() {
-    let mesh_config = parse_config("");
-    let model_file = temp_model_file();
-
-    let resolved = resolve_skippy_config(SkippyConfigResolveRequest {
-        mesh_config: &mesh_config,
-        model_id: "Qwen/Qwen3-0.6B:Q4_K_M",
-        model_path: model_file.path(),
-        model_bytes: 4 * 1024 * 1024 * 1024,
-        allocatable_memory_bytes: None,
-        request_defaults: None,
-        package_generation: None,
-    })
-    .expect("default speculative strategy should resolve");
-
-    assert_eq!(resolved.speculative.strategy, "auto");
-    assert!(!resolved.speculative.native_mtp_enabled);
-    let load_options = resolved
-        .to_model_load_options(SkippyTelemetryOptions::off())
-        .expect("model load options should build");
-    assert!(!load_options.native_mtp_enabled);
-    let stage = resolved
-        .to_stage_config(Some(fake_package_identity(24)), LoadMode::LayerPackage)
-        .expect("stage config should build");
-    assert!(!stage.native_mtp_enabled);
-    let openai = resolved
-        .to_embedded_openai_args(4096, true)
-        .expect("openai args should build");
-    assert!(!openai.native_mtp_enabled);
-}
-
-#[test]
-fn speculative_strategy_auto_uses_package_native_mtp_default() {
-    let mesh_config = parse_config("");
-    let model_file = temp_model_file();
-    let generation = native_mtp_generation();
-
-    let resolved = resolve_skippy_config(SkippyConfigResolveRequest {
-        mesh_config: &mesh_config,
-        model_id: "meshllm/GLM-4.7-Flash-MTP-GGUF",
-        model_path: model_file.path(),
-        model_bytes: 4 * 1024 * 1024 * 1024,
-        allocatable_memory_bytes: None,
-        request_defaults: None,
-        package_generation: Some(&generation),
-    })
-    .expect("package native MTP default should resolve");
-
-    assert_eq!(resolved.speculative.strategy, "auto");
-    assert!(resolved.speculative.native_mtp_enabled);
-    let load_options = resolved
-        .to_model_load_options(SkippyTelemetryOptions::off())
-        .expect("model load options should build");
-    assert!(load_options.native_mtp_enabled);
-    let stage = resolved
-        .to_stage_config(Some(fake_package_identity(24)), LoadMode::LayerPackage)
-        .expect("stage config should build");
-    assert!(stage.native_mtp_enabled);
-    let openai = resolved
-        .to_embedded_openai_args(4096, true)
-        .expect("openai args should build");
-    assert!(openai.native_mtp_enabled);
-}
-
-#[test]
-fn speculative_strategy_native_mtp_rejects_package_without_native_mtp_metadata() {
-    let mesh_config = parse_config(
-        r#"
-[defaults.speculative]
-strategy = "native-mtp-n1"
-"#,
-    );
-    let model_file = temp_model_file();
-    let generation = PackageGenerationInfo {
-        speculative_decoding: None,
-    };
-
-    let error = resolve_skippy_config(SkippyConfigResolveRequest {
-        mesh_config: &mesh_config,
-        model_id: "meshllm/package-without-mtp",
-        model_path: model_file.path(),
-        model_bytes: 4 * 1024 * 1024 * 1024,
-        allocatable_memory_bytes: None,
-        request_defaults: None,
-        package_generation: Some(&generation),
-    })
-    .unwrap_err()
-    .to_string();
-
-    assert!(error.contains("requires package generation metadata advertising native-mtp-n1"));
-}
-
-#[test]
-fn speculative_strategy_disabled_reaches_stage_and_openai_args() {
-    let mesh_config = parse_config(
-        r#"
-[defaults.speculative]
-strategy = "disabled"
-"#,
-    );
-    let model_file = temp_model_file();
-
-    let resolved = resolve_skippy_config(SkippyConfigResolveRequest {
-        mesh_config: &mesh_config,
-        model_id: "Qwen/Qwen3-0.6B:Q4_K_M",
-        model_path: model_file.path(),
-        model_bytes: 4 * 1024 * 1024 * 1024,
-        allocatable_memory_bytes: None,
-        request_defaults: None,
-        package_generation: None,
-    })
-    .expect("disabled speculative strategy should resolve");
-
-    assert_eq!(resolved.speculative.strategy, "disabled");
-    assert!(!resolved.speculative.native_mtp_enabled);
-    let load_options = resolved
-        .to_model_load_options(SkippyTelemetryOptions::off())
-        .expect("model load options should build");
-    assert!(!load_options.native_mtp_enabled);
-    let stage = resolved
-        .to_stage_config(Some(fake_package_identity(24)), LoadMode::LayerPackage)
-        .expect("stage config should build");
-    assert!(!stage.native_mtp_enabled);
-    let openai = resolved
-        .to_embedded_openai_args(4096, true)
-        .expect("openai args should build");
-    assert!(!openai.native_mtp_enabled);
 }
 
 #[test]
